Improve is_empty check on Goal Definitions
This now captures lists of empty goal definitions, such as
`(and (and) (and))`.

Previously such a situation would conclude that the goal is
non-empty (due to the two embedded goals), when in reality
no goal was specified.

The implementation using `.iter().all(...)` works because `all()`
returns `true` on an empty iterator.